I broke things. then fixed it. Changed many ArrayLists into List<T>'s

+++ client/SafeNativeMethods.cs
@@ -20,6 +20,6 @@ namespace myseq
 
         [DllImport("winmm.dll")]
 
-        internal static extern long PlaySound(string lpszName, long hModule, long dwFlags);
+        internal static extern long PlaySound(string lpszName, System.IntPtr hModule, System.IntPtr dwFlags);
     }
 }
